chore: rename and delete unused components (#1016)
