+++ packages/client/hmi-client/src/components/widgets/tera-import-code-button.vue
@@ -8,7 +8,7 @@
 		/>
 		<a :href="urlString" rel="noreferrer noopener">{{ urlString }}</a>
 		<Teleport to="body">
-			<modal v-if="isModalVisible" class="modal" @modal-mask-clicked="isModalVisible = false">
+			<tera-modal v-if="isModalVisible" class="modal" @modal-mask-clicked="isModalVisible = false">
 				<template #header>
 					<h5>
 						Choose file to open from {{ repoOwnerAndName
@@ -32,7 +32,7 @@
 						</li>
 					</ul>
 				</template>
-			</modal>
+			</tera-modal>
 		</Teleport>
 	</main>
 </template>
@@ -40,7 +40,7 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue';
 import Button from 'primevue/button';
-import modal from '@/components/widgets/Modal.vue';
+import TeraModal from '@/components/widgets/tera-modal.vue';
 import { ProjectAssetTypes } from '@/types/Project';
 import { isEmpty } from 'lodash';
 import { getGithubRepositoryContent, getGithubCode } from '@/services/github-import';
