+++ packages/client/hmi-client/src/components/widgets/tera-pdf-embed.vue
