+++ packages/client/hmi-client/src/page/project/components/tera-resource-sidebar.vue
@@ -78,7 +78,7 @@
 			</AccordionTab>
 		</Accordion>
 		<Teleport to="body">
-			<modal
+			<tera-modal
 				v-if="isRemovalModal"
 				@modal-mask-clicked="isRemovalModal = false"
 				class="remove-modal"
@@ -96,7 +96,7 @@
 					<Button label="Remove" class="p-button-danger" @click="removeAsset()" />
 					<Button label="Cancel" class="p-button-secondary" @click="isRemovalModal = false" />
 				</template>
-			</modal>
+			</tera-modal>
 		</Teleport>
 	</nav>
 </template>
@@ -105,7 +105,7 @@
 import { computed, ref } from 'vue';
 import { capitalize, isEmpty, isEqual } from 'lodash';
 import { Tab } from '@/types/common';
-import Modal from '@/components/widgets/Modal.vue';
+import TeraModal from '@/components/widgets/tera-modal.vue';
 import { getAssetIcon } from '@/services/project';
 import Accordion from 'primevue/accordion';
 import AccordionTab from 'primevue/accordiontab';
